+++ src/pages/NwbPage/plugins/Figpack/index.ts
@@ -1,0 +1,29 @@
+import { getHdf5Group } from "@hdf5Interface";
+import { NwbObjectViewPlugin } from "../pluginInterface";
+import FigpackRasterPlotView from "./FigpackRasterPlotView";
+
+export const figpackRasterPlotPlugin: NwbObjectViewPlugin = {
+  name: "FigpackRasterPlot",
+  label: "Figpack Raster",
+  canHandle: async ({
+    nwbUrl,
+    path,
+    secondaryPaths,
+  }: {
+    nwbUrl: string;
+    path: string;
+    secondaryPaths?: string[];
+  }) => {
+    if (secondaryPaths && secondaryPaths.length > 0) return false;
+    const group = await getHdf5Group(nwbUrl, path);
+    if (!group) return false;
+    if (group.attrs["neurodata_type"] === "Units") return true;
+    return false;
+  },
+  component: FigpackRasterPlotView,
+  launchableFromTable: true,
+  requiresWindowDimensions: false,
+  showInMultiView: false,
+};
+
+export default figpackRasterPlotPlugin;